# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 노들섬 홍보 통합 시스템 API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  clerkId: string;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  phone?: string | null;
  role: UserRole;
  /** @nullable */
  organizationId?: number | null;
  createdAt: string;
}