# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 노들섬 홍보 통합 시스템 API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  name?: string;
  phone?: string;
}